requirements.txt
fastapi
uvicorn
pydantic
anthropic
pymongo
python-dotenv
boto3
botocore
python-multipart
bcrypt
passlib[bcrypt]
starlette
itsdangerous
